+++ PalmHill.BlazorChat/Client/WebSocketChatConnection.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.AspNetCore.SignalR.Client;
+using PalmHill.BlazorChat.Client.Models;
 using PalmHill.BlazorChat.Shared.Models;
 using PalmHill.BlazorChat.Shared.Models.WebSocket;
 
@@ -6,15 +7,18 @@ namespace PalmHill.BlazorChat.Client
 {
     public class WebSocketChatConnection
     {
-        public WebSocketChatConnection(Uri chatHubUri)
+        public WebSocketChatConnection(Uri chatHubUri, List<WebSocketChatMessage> webSocketChatMessages)
         {
+            WebSocketChatMessages = webSocketChatMessages;
             HubConnection = new HubConnectionBuilder()
            .WithUrl(chatHubUri)
            .Build();
 
             setupHubConnection();
         }
-
+        public Guid ConversationId { get; } = Guid.NewGuid();
+        public string SystemMessage = string.Empty;
+        public List<WebSocketChatMessage> WebSocketChatMessages { get; }
         public HubConnection HubConnection { get; }
 
         public event EventHandler<WebSocketInferenceString>? OnReceiveInferenceString;
@@ -32,8 +36,9 @@ namespace PalmHill.BlazorChat.Client
             await HubConnection.StopAsync();
         }
 
-        public async Task SendInferenceRequestAsync(InferenceRequest inferenceRequest)
+        public async Task SendInferenceRequestAsync()
         {
+            var inferenceRequest = GetInferenceRequestFromWebsocketMessages();
             await HubConnection.SendAsync("InferenceRequest", inferenceRequest);
         }
 
@@ -41,11 +46,20 @@ namespace PalmHill.BlazorChat.Client
         {
             HubConnection.On<WebSocketInferenceString>("ReceiveInferenceString", (inferenceString) =>
             {
+                var lastPrompt = WebSocketChatMessages.SingleOrDefault(cm => cm.Id == inferenceString.MessageId);
+                lastPrompt?.AddResponseString(inferenceString.InferenceString);
                 OnReceiveInferenceString?.Invoke(this, inferenceString);
             });
 
             HubConnection.On<WebSocketInferenceStatusUpdate>("InferenceStatusUpdate", (statusUpdate) =>
             {
+                var lastPrompt = WebSocketChatMessages.Single(cm => cm.Id == statusUpdate.MessageId);
+
+                if (statusUpdate.IsComplete)
+                {
+                    lastPrompt.CompleteResponse(statusUpdate.Success ?? false);
+                }
+
                 OnInferenceStatusUpdate?.Invoke(this, statusUpdate);
             });
 
@@ -53,6 +67,34 @@ namespace PalmHill.BlazorChat.Client
             {
                 OnAttachmentStatusUpdate?.Invoke(this, attachmentInfo);
             });
+        }
+
+        private InferenceRequest GetInferenceRequestFromWebsocketMessages()
+        {
+            var chatConversation = new InferenceRequest();
+            chatConversation.Id = ConversationId;
+            chatConversation.SystemMessage = SystemMessage;
+
+            foreach (var promptAndResponse in WebSocketChatMessages)
+            {
+                var userMessage = new ChatMessage();
+                userMessage.Message = promptAndResponse.Prompt;
+                userMessage.Id = promptAndResponse.Id;
+                userMessage.Role = ChatMessageRole.User;
+                chatConversation.ChatMessages.Add(userMessage);
+
+                if (promptAndResponse.IsComplete && promptAndResponse.Success == true)
+                {
+                    var modelMessage = new ChatMessage();
+                    modelMessage.Message = promptAndResponse.Resonse;
+                    modelMessage.Role = ChatMessageRole.Assistant;
+                    chatConversation.ChatMessages.Add(modelMessage);
+
+                }
+
+            }
+
+            return chatConversation;
         }
     }
 }
